refactor: (재고 유효성 검증 - 재료 사용하여 주스 제조) 로직을 타입으로 뺐으나 다시 JuiceMaker 내의 메서드로 변경
