+++ JuiceMaker/JuiceMaker/MakeJuice/View/MakeJuiceViewController.swift
@@ -37,7 +37,7 @@ class MakeJuiceViewController: UIViewController {
             try viewModel.order(juice)
             let alert = CompleteMakingAlert()
             self.present(alert.alertController(), animated: true, completion: nil)
-        } catch JuiceManufacturerError.soldOut {
+        } catch JuiceMaker.JuiceMakerError.soldOut {
             let alert = SoldOutAlert()
             self.present(alert.alertController(), animated: true, completion: nil)
         } catch {
